+++ R/nano.R
@@ -99,6 +99,12 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
     r <- dial(socket, url = dial, autostart = autostart)
     if (r == 0L) {
       nano[["dialer"]] <- attr(socket, "dialer")
+      nano[["dialer_getopt"]] <- function(type = c("bool", "int", "ms", "size",
+                                                   "string", "uint64"),
+                                          opt) lapply(.subset2(nano, "dialer"),
+                                                      getopt,
+                                                      type = type,
+                                                      opt = opt)
       nano[["dialer_setopt"]] <- function(type = c("bool", "int", "ms", "size",
                                                    "string", "uint64"),
                                           opt,
@@ -119,6 +125,12 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
     r <- listen(socket, url = listen, autostart = autostart)
     if (r == 0L) {
       nano[["listener"]] <- attr(socket, "listener")
+      nano[["listener_getopt"]] <- function(type = c("bool", "int", "ms", "size",
+                                                     "string", "uint64"),
+                                            opt) lapply(.subset2(nano, "listener"),
+                                                        getopt,
+                                                        type = type,
+                                                        opt = opt)
       nano[["listener_setopt"]] <- function(type = c("bool", "int", "ms", "size",
                                                      "string", "uint64"),
                                             opt,
@@ -141,6 +153,12 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
     r <- dial(socket, url = url, autostart = autostart)
     if (r == 0L) {
       nano[["dialer"]] <- attr(socket, "dialer")
+      nano[["dialer_getopt"]] <- function(type = c("bool", "int", "ms", "size",
+                                                   "string", "uint64"),
+                                          opt) lapply(.subset2(nano, "dialer"),
+                                                      getopt,
+                                                      type = type,
+                                                      opt = opt)
       nano[["dialer_setopt"]] <- function(type = c("bool", "int", "ms", "size",
                                                    "string", "uint64"),
                                           opt,
@@ -162,6 +180,12 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
     r <- listen(socket, url = url, autostart = autostart)
     if (r == 0L) {
       nano[["listener"]] <- attr(socket, "listener")
+      nano[["listener_getopt"]] <- function(type = c("bool", "int", "ms", "size",
+                                                     "string", "uint64"),
+                                            opt) lapply(.subset2(nano, "listener"),
+                                                        getopt,
+                                                        type = type,
+                                                        opt = opt)
       nano[["listener_setopt"]] <- function(type = c("bool", "int", "ms", "size",
                                                      "string", "uint64"),
                                             opt,
@@ -178,6 +202,11 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
     }
     invisible(r)
   }
+  nano[["getopt"]] <- function(type = c("bool", "int", "ms", "size",
+                                        "string", "uint64"),
+                               opt) getopt(socket,
+                                           type = type,
+                                           opt = opt)
   nano[["recv"]] <- function(mode = c("serial", "character", "complex", "double",
                                       "integer", "logical", "numeric", "raw"),
                              block = NULL,