+++ src/nanonext.h
@@ -128,6 +128,7 @@ extern SEXP rnng_device(SEXP, SEXP);
 extern SEXP rnng_dial(SEXP, SEXP, SEXP);
 extern SEXP rnng_dialer_close(SEXP);
 extern SEXP rnng_dialer_start(SEXP, SEXP);
+extern SEXP rnng_get_opt(SEXP, SEXP, SEXP);
 extern SEXP rnng_listen(SEXP, SEXP, SEXP);
 extern SEXP rnng_listener_close(SEXP);
 extern SEXP rnng_listener_start(SEXP);